Make the stream compatible with more browsers
