+++ package.json
@@ -33,7 +33,8 @@
   },
   "dependencies": {
     "@metamask/utils": "^3.0.1",
-    "readable-stream": "2.3.3"
+    "readable-stream": "2.3.3",
+    "webextension-polyfill": "^0.10.0"
   },
   "devDependencies": {
     "@jest-runner/electron": "^3.0.1",
@@ -47,6 +48,7 @@
     "@types/jest": "^26.0.13",
     "@types/node": "^18.0.0",
     "@types/readable-stream": "^2.3.9",
+    "@types/webextension-polyfill": "^0.9.2",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
     "browserify": "^17.0.0",
