+++ src/browser.test.ts
@@ -1,5 +1,10 @@
 import * as PostMessageStream from './browser';
 
+// `BrowserRuntimePostMessageStream` imports `webextension-polyfill` which
+// throws an error if it's not in a browser environment. We need to mock it
+// here to prevent that error.
+jest.mock('webextension-polyfill', () => undefined);
+
 describe('post-message-stream', () => {
   describe('browser exports', () => {
     const expectedExports = [
@@ -7,7 +12,7 @@ describe('post-message-stream', () => {
       'WindowPostMessageStream',
       'WebWorkerPostMessageStream',
       'WebWorkerParentPostMessageStream',
-      'RuntimePostMessageStream',
+      'BrowserRuntimePostMessageStream',
     ];
 
     it('package has expected exports', () => {
