+++ src/browser.ts
@@ -2,6 +2,6 @@
 export * from './window/WindowPostMessageStream';
 export * from './WebWorker/WebWorkerPostMessageStream';
 export * from './WebWorker/WebWorkerParentPostMessageStream';
-export * from './runtime/RuntimePostMessageStream';
+export * from './runtime/BrowserRuntimePostMessageStream';
 export * from './BasePostMessageStream';
 export { StreamData, StreamMessage } from './utils';
