+++ src/index.test.ts
@@ -1,5 +1,10 @@
 import * as PostMessageStream from '.';
 
+// `BrowserRuntimePostMessageStream` imports `webextension-polyfill` which
+// throws an error if it's not in a browser environment. We need to mock it
+// here to prevent that error.
+jest.mock('webextension-polyfill', () => undefined);
+
 describe('post-message-stream', () => {
   describe('exports', () => {
     const expectedExports = [
@@ -11,7 +16,7 @@ describe('post-message-stream', () => {
       'ProcessMessageStream',
       'ThreadParentMessageStream',
       'ThreadMessageStream',
-      'RuntimePostMessageStream',
+      'BrowserRuntimePostMessageStream',
     ];
 
     it('package has expected exports', () => {
