+++ src/index.ts
@@ -5,6 +5,6 @@ export * from './node-process/ProcessParentMessageStream';
 export * from './node-process/ProcessMessageStream';
 export * from './node-thread/ThreadParentMessageStream';
 export * from './node-thread/ThreadMessageStream';
-export * from './runtime/RuntimePostMessageStream';
+export * from './runtime/BrowserRuntimePostMessageStream';
 export * from './BasePostMessageStream';
 export { StreamData, StreamMessage } from './utils';
